+++ cmd/spiffe-helper/config/config_posix.go
@@ -1,4 +1,4 @@
-package main
+package config
 
 func validateOSConfig(*Config) error {
 	return nil
