+++ cmd/spiffe-helper/config/config_test.go
@@ -1,4 +1,4 @@
-package main
+package config
 
 import (
 	"os"
@@ -11,7 +11,7 @@ import (
 )
 
 func TestParseConfig(t *testing.T) {
-	c, err := ParseConfig("../../test/fixture/config/helper.conf")
+	c, err := ParseConfig("../../../test/fixture/config/helper.conf")
 
 	assert.NoError(t, err)
